pyproject.toml
[tool.poetry]
name = "grant-back"
version = "0.1.0"
description = ""
authors = ["AliRN <ali_ramazanov_2000@mail.ru>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.0.1"
matplotlib = "^3.8.2"
fastapi = { extras = ["standard"], version = "^0.115.4" }
uvicorn = { extras = ["standard"], version = "^0.32.0" }
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
aiofiles = "^24.1.0"
sympy = "^1.14.0"
scipy = "^1.15.3"
bidict = "^0.23.1"
starlette-context = "^0.4.0"
pandas = "^2.2.3"
scikit-learn = "^1.6.1"
openpyxl = "^3.1.5"
autograd = "^1.6.2"



[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pylint = "^3.1.0"
cookiecutter = "^2.6.0"


[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
